Move SignOunButton component to settings folder
