+++ src/components/settings/signOutButton/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { NavigationActions } from 'react-navigation';
-import { removePassphraseFromKeyChain } from '../../utilities/passphrase';
-import { IconButton } from '../toolBox/button';
-import { colors } from '../../constants/styleGuide';
-import withTheme from '../withTheme';
+import { removePassphraseFromKeyChain } from '../../../utilities/passphrase';
+import { IconButton } from '../../toolBox/button';
+import { colors } from '../../../constants/styleGuide';
+import withTheme from '../../withTheme';
 import getStyles from './styles';
 
 class SignOutButton extends React.Component {
