+++ src/components/settings/signOutButton/styles.js
@@ -1,4 +1,4 @@
-import { themes, colors } from '../../constants/styleGuide';
+import { themes, colors } from '../../../constants/styleGuide';
 
 export default () => ({
   common: {
